Initial Commit
SQLite can do things like ArcMap's Summary Statistics tool MUCH faster.
This module facilitates exporting a table from ArcGIS (for example, an
attribute table of a shapefile or a table in a file geodatabase) into a
SQLite table as well as converting back the results of a query into an
ArGIS table format.